Revert "up to ac17fbe plus log2d bootlog"
This reverts commit 4ac439147de7cef481f3f9dee633bf5ac7a0d1e6.

+++ src/build-linux.sh
@@ -1,4 +1,19 @@
 #!/bin/sh
+# ABOUT
+# Build shell script for Artisan Linux builds
+#
+# LICENSE
+# This program or module is free software: you can redistribute it and/or
+# modify it under the terms of the GNU General Public License as published
+# by the Free Software Foundation, either version 2 of the License, or
+# version 3 of the License, or (at your option) any later versison. It is
+# provided for educational purposes and is distributed in the hope that
+# it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
+# the GNU General Public License for more details.
+#
+# AUTHOR
+# Dave Baxter, Marko Luther 2023
 
 #set -ex
 set -e  # reduced logging
@@ -13,12 +28,19 @@ if [ ! -z $APPVEYOR ]; then
 elif [ -d /usr/lib/python3/dist-packages/PyQt5 ]; then
     # ARM builds
     export PYTHON_PATH=`python3 -m site --user-site`
+    export PYTHONSITEPKGS=`python3 -m site --user-site`
     export QT_PATH=/usr/share/qt5
+    export QT_SRC_PATH=/usr/share/Qt/6.4/gcc_64
+    export PYLUPDATE=./pylupdate6pro.py
 else
     # Other builds
     export PYTHON_PATH=`python3 -m site --user-site`
     export QT_PATH=$PYTHON_PATH/PyQt5/Qt
 fi
+
+echo "************* build derived files **************"
+./build-derived.sh linux  #generate the derived files
+if [ $? -ne 0 ]; then echo "Failed in build-derived.sh"; exit $?; else (echo "** Finished build-derived.sh"); fi
 
 rm -rf build
 rm -rf dist
@@ -45,12 +67,16 @@ mkdir dist/translations
 for lan in ar, da, de, en, el, es, fa, fi, fr, gd, he, hu, id, it, ja, ko, lv, nl, no, pl, pt_BR, pt, sk, sv, th, tr, uk, vi, zh_CN, zh_TW; do
      QTBASE_FILE=$QT_PATH/translations/qtbase_${lan}.qm
      QT_FILE=$QT_PATH/translations/qt_${lan}.qm
+#     QTCONNECTIVITY_FILE=$QT_PATH/translations/qtconnectivity_${lan}.qm
      if [ -e ${QTBASE_FILE} ]
           then cp ${QTBASE_FILE} dist/translations
      fi
      if [ -e ${QT_FILE} ]
           then cp ${QT_FILE} dist/translations
      fi
+#     if [ -e ${QTCONNECTIVITY_FILE} ]
+#          then cp ${QTCONNECTIVITY_FILE} dist/translations
+#     fi
 done
 
 cp translations/*.qm dist/translations
@@ -107,7 +133,8 @@ cp -R includes/Icons/* dist/Icons
 
 mkdir dist/yoctopuce
 mkdir dist/yoctopuce/cdll
-cp $PYTHON_PATH/yoctopuce/cdll/*64.so dist/yoctopuce/cdll
+# dave cp $PYTHON_PATH/yoctopuce/cdll/*64.so dist/yoctopuce/cdll
+cp $PYTHONSITEPKGS/yoctopuce/cdll/*64.so dist/yoctopuce/cdll
 
 cp /usr/lib/libsnap7.so dist
